Upgrading jetty from 9.2.18 to 9.4.8
One potential breaking change - it has support for TLS1.2 alone. Meaning it disables (by default) all cipher suites that matches ^.*_(MD5|SHA|SHA1)$ which are the only ones which work with TLS 1.0 and TLS1.1. So if any of the client uses TLS1.1 or TLS1.0 then the communication will fail. Need to test if this is an issue for us or not.
This version fixes some known issues with JDK9
fixes a class loader leak issue on the version we were on
bunch of bug fixes and a few feature enhancements